attendance_webapp/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
Flask-Bcrypt>=1.0
Flask-WTF>=1.0
python-dotenv>=1.0